% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectSubtours.R
\name{detectSubtours}
\alias{detectSubtours}
\title{Detect subtours in a given traveling salesman problem (TSP) route.}
\usage{
detectSubtours(x_tsp, idxs)
}
\arguments{
\item{x_tsp}{The matrix representing the tour's tracks with two columns.}

\item{idxs}{The distance matrix of the cities.}
}
\value{
A list of all possible subtours found in the TSP route.
}
\description{
Detect subtours in a given traveling salesman problem (TSP) route.
}
\examples{
detectSubtours(x_tsp, idxs)
}
