% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_TSP.R
\name{global_TSP}
\alias{global_TSP}
\title{Find the shortest path in a traveling salesman problem (TSP) with the first and last destinations fixed.}
\usage{
global_TSP(cities)
}
\arguments{
\item{position_windows}{A vector-based matrix representing the characteristic matrix of each text window. Each row has 300 dimensions.}
}
\value{
A list containing the route and the distance of the shortest path.
}
\description{
Find the shortest path in a traveling salesman problem (TSP) with the first and last destinations fixed.
}
\examples{
global_TSP(position_windows)
}
