% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinVolEllipse.R
\name{MinVolEllipse}
\alias{MinVolEllipse}
\title{Inner function use to calculate the shortest path}
\usage{
MinVolEllipse(P, tolerance)
}
\arguments{
\item{P}{The matrix}

\item{tolerance}{The tolerance of the calcuatoion precision}
}
\value{
The matrix and value required for our further calculation
}
\description{
Inner function use to calculate the shortest path
}
\examples{
MinVolEllipse(P,1e-5)
}
