% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectSubtours.R
\name{detectSubtours}
\alias{detectSubtours}
\title{Subtours detection function}
\usage{
detectSubtours(x_tsp, idxs)
}
\arguments{
\item{x_tsp}{The tours track matrix with two columns}

\item{idxs}{The distance matrix of the cities}
}
\value{
All possible subtour in our route results
}
\description{
Subtours detection function
}
\examples{
detectSubtours(x_tsp, idxs)
}
