% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_the_data.R
\name{get_the_data}
\alias{get_the_data}
\title{Calculate the characteristics summary of speed, volume, and circuitousness for the given text corpus.}
\usage{
get_the_data(
  path_for_corpus,
  path_for_emb,
  own_emb = FALSE,
  type_of_window = "sentence",
  window_length = 250,
  include_extension_speed = FALSE,
  include_extension_volume = FALSE
)
}
\arguments{
\item{path_for_corpus}{The path of your corpus. It should be in CSV format with a "text" column.}

\item{path_for_emb}{If own_emb=FALSE, it means you are using the provided embedding. In this case, this parameter should be the path to the downloaded RData file (link: \url{https://drive.google.com/uc?export=download&id=1Nq--lnyG_9cLdjcjVPe4tVl3nqyIm8WT}). If own_emb=TRUE, you are using your own embedding. Please prepare the embedding (CSV file) in a standard format. You can download a sample embedding with the standard format from this link \url{https://drive.google.com/uc?export=download&id=1koC_c7U2vXJQyl1jAtUa8W7holN1WmLf}.}

\item{own_emb}{Default=FALSE. Set to TRUE to use your own embedding.}

\item{type_of_window}{Default="sentence". There are two alternative values: "sentence" or "length." Use "length" if you want to divide the text by a specified word length. "sentence" ensures that the function doesn't break sentences in the middle but finishes them.}

\item{window_length}{Integer. When type_of_window="length," it divides the text into windows of this specified word length. You can set the length of windows to 1 and type_of_window to "length" to separate the text into individual sentences.}

\item{include_extension_speed}{Include additional result or not. Default=False. (If True, program execution may be slower)}

\item{include_extension_volume}{Include additional volume or not. Default=False. (If True, program execution may be slower)}
}
\value{
A dataset containing the calculated speed, volume, and circuitousness of the text.
}
\description{
Calculate the characteristics summary of speed, volume, and circuitousness for the given text corpus.
}
\examples{
library(rStoryShape)
path_for_emb <- "data/ft_model_test.Rdata"
path_for_corpus <- "data/corpus.Rdata"
result <- get_the_data(path_for_corpus, path_for_emb, own_emb=FALSE, type_of_window = "length", window_length = 10)
print(result)


}
